+++ SampleCode/Android/app/src/main/java/com/netease/meetinglib/demo/MeetingApplication.java
@@ -8,6 +8,7 @@ package com.netease.meetinglib.demo;
 import android.app.Application;
 
 import com.netease.meetinglib.demo.log.LogUtil;
+import com.netease.meetinglib.demo.nim.NIMInitializer;
 
 public class MeetingApplication extends Application {
     private static MeetingApplication instance;
@@ -17,6 +18,7 @@ public class MeetingApplication extends Application {
         super.onCreate();
         instance = this;
         LogUtil.init(this);
+        NIMInitializer.getInstance().startInitialize(this);
         SdkAuthenticator.getInstance().initialize(this);
         SdkInitializer.getInstance().startInitialize(this);
     }