+++ SampleCode/iOS/NEMeetingDemo/Section/StartMeetingVC.m
@@ -29,6 +29,7 @@
 @property (nonatomic, readonly) BOOL disableChat;
 @property (nonatomic, readonly) BOOL disableInvite;
 @property (nonatomic, readonly) BOOL disableMinimize;
+@property (nonatomic, readonly) BOOL disableGallery;
 
 @property (nonatomic, strong) NSMutableArray <NEMeetingMenuItem *> *menuItems;
 
@@ -60,7 +61,10 @@
                                               @"入会时关闭邀请菜单",
                                               @"入会时隐藏最小化",
                                               @"使用个人会议号",
-                                              @"使用默认会议设置"]];
+                                              @"使用默认会议设置",
+                                              @"入会时关闭画廊模式",
+                                              @"仅显示会议ID长号",
+                                              @"仅显示会议ID短号"]];
     [_settingCheckBox setItemSelected:YES index:2];
     _settingCheckBox.delegate = self;
 }
@@ -77,10 +81,12 @@
         options.noVideo = ![self openVideoWhenJoinMeeting];
         options.noAudio = ![self openMicWhenJoinMeeting];
         options.showMeetingTime = [self showMeetingTime];
+        options.meetingIdDisplayOption = [self meetingIdDisplayOption];
         options.noInvite = [self disableInvite];
         options.noChat = [self disableChat];
         options.noMinimize = [self disableMinimize];
         options.injectedMoreMenuItems = _menuItems;
+        options.noGallery = [self disableGallery];
     }
     
     WEAK_SELF(weakSelf);
@@ -190,6 +196,19 @@
 
 - (BOOL)useDefaultConfig {
     return [_settingCheckBox getItemSelectedAtIndex:4];
+}
+
+- (BOOL)disableGallery {
+    return [_settingCheckBox getItemSelectedAtIndex:5];
+}
+
+- (NEMeetingIdDisplayOption) meetingIdDisplayOption {
+    if ([_settingCheckBox getItemSelectedAtIndex:6]) {
+        return DISPLAY_LONG_ID_ONLY;
+    } else if ([_settingCheckBox getItemSelectedAtIndex:7]) {
+        return DISPLAY_SHORT_ID_ONLY;
+    }
+    return DISPLAY_ALL;
 }
 
 @end